Initial commit
Note: I didnt start commiting at the beginning of this project. but
at this current state, the project has made considerable progress.
the apps homepage, offline hymns list page and settings have been
implemented but the features in the settings are not working yet.
the app search hymns functionality is woring.